+++ shoppingOnline/src/main/java/com/roninEngineerF02/shoppingOnline/exception/ProductNotFoundException.java
@@ -1,7 +1,7 @@
 package com.roninEngineerF02.shoppingOnline.exception;
 
 public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Integer productId) {
+    public ProductNotFoundException(Long productId) {
         super("Product not found with ID: " + productId);
     }
 }
