+++ shoppingOnline/src/main/java/com/roninEngineerF02/shoppingOnline/repository/CartItemRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
-    Optional<CartItem> findByCart_IdAndProduct_Id(Integer cartId, Integer productId);
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+    Optional<CartItem> findByCart_IdAndProduct_Id(Long cartId, Long productId);
 }
